Two AI currently play for 80 turns for debug purposes.
Confirmed that placing random ships, shooting and registering if a shot is a hit, or miss and if hit if sunk are all working again.

Made significant changes to logic mainly to utilise the GuessCell and MutableMatrix classes. Also refactored allot of functions to correct classes.

Made significant changes to UI to make it much simpler.